[#15]; refactor: 컨트롤러 패키지명 변경 auth 기능에 맞게 컨트롤러 패키지명 변경
